Refactor foreign configurations off controller class
