+++ pystorcli2/controller/__init__.py
@@ -11,7 +11,6 @@ from .. import StorCLI
 from .. import common
 from .. import enclosure
 from .. import virtualdrive
-from .. import foreignconfig
 from .. import exc
 
 from datetime import datetime
@@ -37,11 +36,11 @@ class Controller(object):
         metrics (:obj:ControllerMetrics): controller metrics
         vds (list of :obj:virtualdrive.VirtualDrives): controller virtual drives
         encls (:obj:enclosure.Enclosures): controller enclosures
-        fcs (:obj:foreignconfig:ForeignConfigurations): controller foreign configurations
         autorebuild (dict): current auto rebuild state (also setter)
         foreignautoimport (dict): imports foreign configuration automatically at boot (also setter)
         patrolread (dict): current patrol read settings (also setter)
         cc (dict): current patrol read settings (also setter)
+        has_foreign_configurations (bool): true if controller has foreign configurations
 
     Methods:
         create_vd (:obj:VirtualDrive): create virtual drive
@@ -52,6 +51,8 @@ class Controller(object):
         patrolread_stop (dict): stops patrol read if running on controller
         patrolread_running (bool): check if patrol read is running on controller
         set_cc (dict): configures consistency check mode and start time
+        import_foreign_configurations (dict): imports the foreign configurations on controller
+        delete_foreign_configurations (dict): deletes the foreign configuration on controller
 
     TODO:
         Implement missing methods:
@@ -126,12 +127,6 @@ class Controller(object):
         """(:obj:enclosure.Enclosures): controller enclosures
         """
         return enclosure.Enclosures(ctl_id=self._ctl_id, binary=self._binary)
-
-    @property
-    def fcs(self):
-        """(:obj:foreignconfig.ForeignConfigurations): controller foreign configurations
-        """
-        return foreignconfig.ForeignConfigurations(ctl_id=self._ctl_id, binary=self._binary)
 
     @property
     def drives_ids(self) -> List[str]:
@@ -423,6 +418,53 @@ class Controller(object):
             args.append('starttime="{0}"'.format(starttime))
 
         return common.response_setter(self._run(args))
+
+    @property
+    def has_foreign_configurations(self) -> bool:
+        """(bool): true if controller has foreign configurations
+        """
+        args = [
+            '/fall',
+            'show'
+        ]
+
+        try:
+            fcs = common.response_data(self._run(args))['Total foreign Drive Groups']
+            if fcs > 0:
+                return True
+        except KeyError:
+            pass
+        return False
+
+    def delete_foreign_configurations(self, securitykey: Optional[str] = None):
+        """Deletes foreign configurations
+
+        Returns:
+            (dict): response cmd data
+        """
+        args = [
+            '/fall',
+            'del'
+        ]
+
+        if securitykey:
+            args.append(f'securitykey={securitykey}')
+        return common.response_cmd(self._run(args))
+
+    def import_foreign_configurations(self, securitykey: Optional[str] = None):
+        """Imports foreign configurations
+
+        Returns:
+            (dict): response cmd data
+        """
+        args = [
+            '/fall',
+            'del'
+        ]
+        if securitykey:
+            args.append(f'securitykey={securitykey}')
+        return common.response_cmd(self._run(args))
+
 
 
 class Controllers(object):
